requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.20

# Resume processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# RSS feed parsing
feedparser>=6.0.10

# HTML parsing and text processing
beautifulsoup4>=4.12.0
html5lib>=1.1

# Embeddings and ML
numpy>=1.24.0
sentence-transformers>=2.2.0
torch>=2.0.0
scikit-learn>=1.3.0

# CORS is handled by FastAPI's CORSMiddleware (no additional package needed)

